Update to new calf internal & data objects
This commit contains a near-complete rewrite of
the package, including a new calf internal function
and new datasets. The new computation greatly
simplifies the code for better understanding and
future development. The new data matches the data
to be published in the bioninformatics publication.

For now, wrappers other than 'calf' have been tabled
as they are not necessary to recreate the examples
in the paper.

This commit still needs more work on documentation
before submitting to CRAN. Additionally, the package
would benefit from the user manual being added as
a vignette. Additional vignettes will be added that
reproduce the paper results using the updated
package.